Branch overlapping fixed
 - Removing transition - Animation no necessary on background change
 - Refactoring - Code for generate branches moved to different file
 - Refactoring - Magic numbers were assigned to variables
 - Name refactoring from 'utils.diagram' to 'utils'
