+++ src/client/app/flogo.flows.detail.diagram/components/diagram.component.less
@@ -490,7 +490,6 @@
       margin: 0;
       bottom: -10px;
       right: 2px;
-      transition: all 250ms linear;
       width: 95px;
 
       .img-left,
